Try to restructure code in the handlers package
Give files sensible names and break it out a bit. I want to make it more
discoverable for people that are new to Go.
