+++ pkg/handlers/handlers.go
@@ -8,29 +8,12 @@ import (
 	"github.com/alphagov/performance-datastore/pkg/dataset"
 	"github.com/alphagov/performance-datastore/pkg/validation"
 	"github.com/go-martini/martini"
-	"github.com/quipo/statsd"
-	"gopkg.in/unrolled/render.v1"
 	"io/ioutil"
 	"net/http"
 	"reflect"
 	"strings"
 	"time"
 )
-
-type ErrorInfo struct {
-	Id     string   `json:"id,omitempty"`
-	HREF   string   `json:"href,omitempty"`
-	Status string   `json:"status,omitempty"`
-	Code   string   `json:"code,omitempty"`
-	Title  string   `json:"title,omitempty"`
-	Detail string   `json:"detail,omitempty"`
-	Links  []string `json:"links,omitempty"`
-	Path   string   `json:"path,omitempty"`
-}
-
-type errorResponse struct {
-	Errors []*ErrorInfo `json:"errors"`
-}
 
 type WarningResponse struct {
 	Data    interface{}
@@ -53,25 +36,6 @@ func NewHandler(maxGzipBody int, logger *logrus.Logger) http.Handler {
 	m.Post("/data/:data_group/:data_type", CreateHandler)
 	m.Put("/data/:data_group/:data_type", UpdateHandler)
 	return m
-}
-
-var (
-	// DataSetStorage is the application global for talking to persistent storage
-	// It is like this to allow test implementations to be injected.
-	DataSetStorage dataset.DataSetStorage
-
-	// ConfigAPIClient allows the client to be injected for testing purposes
-	ConfigAPIClient config_api.Client
-
-	// StatsdClient allows the statsd implementation to be injected for testing purposes
-	StatsdClient statsd.Statsd
-
-	renderer = render.New(render.Options{})
-)
-
-func MethodNotAllowedHandler(w http.ResponseWriter, r *http.Request) {
-	renderError(w, http.StatusMethodNotAllowed,
-		"Method "+r.Method+" not allowed for <"+r.URL.RequestURI()+">")
 }
 
 // DataTypeHandler is responsible for serving data type meta data
@@ -262,22 +226,6 @@ func extractBearerToken(dataSet dataset.DataSet, authorization string) (token st
 
 	token = authorization[len(prefix):]
 	return token, token == dataSet.BearerToken()
-}
-
-func renderError(w http.ResponseWriter, status int, errorString string) {
-	renderer.JSON(w, status, &errorResponse{Errors: []*ErrorInfo{&ErrorInfo{Detail: errorString}}})
-}
-
-func NewStatsDClient(host, prefix string) *statsd.StatsdClient {
-	statsdClient := statsd.NewStatsdClient(host, prefix)
-	statsdClient.CreateSocket()
-
-	return statsdClient
-}
-
-func statsDTiming(label string, start, end time.Time) {
-	StatsdClient.Timing("time."+label,
-		int64(end.Sub(start)/time.Millisecond))
 }
 
 func fetchDataMetaData(dataGroup string, dataType string) (*config_api.DataSetMetaData, error) {
